moved form array methods into helper file
